unmovable window by default

+++ src/options.h
@@ -18,10 +18,10 @@ public:
     struct WindowProperties
     {
         std::string title;
-        uint16_t left;
-        uint16_t top;
-        uint16_t width;
-        uint16_t height;
+        uint32_t left;
+        uint32_t top;
+        uint32_t width;
+        uint32_t height;
         bool fullscreen;
     };
